#56: Moved SwiftUI stuff to new file
